Refactored 'movie list' component
